+++ src/player/main/sp/it/pl/main/AppActions.kt
@@ -150,7 +150,11 @@ class AppActions: GlobalSubConfigDelegator("Shortcuts") {
 
    @IsAction(name = "Open settings", desc = "Opens application settings.")
    fun openSettings() {
-      APP.widgetManager.widgets.use<ConfiguringFeature>(NO_LAYOUT) { it.configure(APP.configuration) }
+      openSettings(null)
+   }
+
+   fun openSettings(groupToSelect: String?) {
+      APP.widgetManager.widgets.use<ConfiguringFeature>(NO_LAYOUT) { it.configure(APP.configuration, groupToSelect) }
    }
 
    @IsAction(name = "Open app actions", desc = "Actions specific to whole application.")